angular

+++ src/app/task-list-component/task-list-component.component.ts
@@ -1,37 +1,57 @@
 import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { TaskService } from '../task.service';
-import { Router } from '@angular/router';
+import { TaskService, task } from '../task.service';
+import { Router, RouterModule } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
+import { SearchPipe } from '../search.pipe';
+import { UserService } from '../user.service';
 
 @Component({
   selector: 'app-task-list-component',
   standalone: true,
-  imports: [FormsModule,CommonModule],
+  imports: [FormsModule,CommonModule,RouterModule,SearchPipe],
   templateUrl: './task-list-component.component.html',
   styleUrl: './task-list-component.component.css'
 })
 export class TaskListComponentComponent implements OnInit {
+  searchTerm: string = '';
 
-  task:any[] =[];
+  task:task[] =[];
+User: any;
+username:any[]=[];
 
-  constructor(private taskService : TaskService){
+
+  constructor(private taskService : TaskService, private router: Router, private toastr: ToastrService, private userService: UserService){
 
   }
 
+  
+
   ngOnInit(): void {
-   this.taskService.getTask().subscribe(a=>
-    this.task = a
-   )
+   this.loadTask();
+   
+  }
+
+  loadTask():void{
+    this.taskService.getTasks().subscribe(a=>
+      this.task = a
+     )
   }
 
   OnDelete(taskId: number){
+    if(confirm('Do you want to delete?'))
     this.taskService.deleteTask(taskId).subscribe(data =>{
-      alert('task is Delete Successfully!...');
-      this.taskService.getTask().subscribe(a=>
-        this.task = a
-      )
+      this.toastr.success('Task is deleted', "Deleted",{
+        timeOut:8000,
+        closeButton:true
+      });
+      this.loadTask();
+      
     })
   }
 
+ onEdit(taskId:number){
+  this.router.navigate(['/edit',taskId])
+ }
 }
